apps/api/requirements.txt
fastapi==0.100.0
langchain==0.0.232
uvicorn==0.22.0
python-dotenv==1.0.0
sqlalchemy==2.0.18
psycopg2-binary==2.9.6
pinecone-client==2.2.2